+++ admin_sections/file_management.php
@@ -14,6 +14,7 @@
             <div>Filename</div>
             <div>Type</div>
             <div>Size</div>
+            <div>Downloads</div>
             <div>Uploaded</div>
             <div>Expires</div>
             <div>Preview</div>
@@ -21,13 +22,15 @@
         </div>
 
         <?php foreach ($allFiles as $file): ?>
+            <?php $fileIcon = get_file_icon($file['filetype'], $file['original_name'] ?? ''); ?>
             <div class="file-row-file-management">
                 <div><?= sanitize_data($file['username'] ?? 'Guest') ?></div>
-                <div><?= sanitize_data($file['original_name']) ?></div>
+                <div><span class="file-icon"><?= (str_starts_with($fileIcon, 'http') ? '<img src="' . sanitize_data($fileIcon) . '" alt="" class="file-icon-img">' : $fileIcon) ?></span> <?= sanitize_data($file['original_name']) ?></div>
                 <div title="<?= sanitize_data($file['filetype']) ?>">
                     <?= sanitize_data(get_friendly_filetype($file['filetype'])) ?>
                 </div>
                 <div><?= format_filesize($file['filesize']) ?></div>
+                <div><?= (int) ($file['download_count'] ?? 0) ?></div>
                 <div><span class="utc-datetime" data-utc="<?= sanitize_data($file['upload_date']) ?>"></span></div>
                 <div>
                     <?= $file['expiry_date']